==> 03_machine-learning.Rmd <==
# Predicting tuber yield category {#Chapter-Modeling}

## Objective

***
The objective of this chapter is to develop, evaluate and compare the performance of some machine learning algorithms (k-nearest neighbors, random forest and support vector machine - [package caret](https://topepo.github.io/caret/index.html)) in predicting tuber yield categories using clr coordinates. We use the previous chapter (chapter \@ref(Chapter-Clustering)) tidded data file `dfml.csv` which contains clr coordinates, maturity classes and the yield two categorical variable created using the 65^th percentile for each cultivar. We use `accuracy` as models quality meeasure. We run a _Chi-square homogenity test_ to compare the best model (with highest accuracy) with a _random classifier_ consisting of an equal distribution of 50% successful and 50% unsuccessful cases.   

We finally compute Euclidean distance as the measure of the multivariate distance between an observation and the closest true negative. By true negative or nutritionally balanced specimens, we mean the samples correctely predicted by the best model as high yielders in the training set. The training and testing sets are stored for the next chapter (preturbation concept - Chapter \@ref(Chapter-Perturbation)).

***

## Useful libraries

The `tidyverse` package is always needed for data easy manipulation and visualization, and then `extrafont` to make changes in graphs as demanded for the article. The particularly useful packages are `caret` and [kknn](https://www.rdocumentation.org/packages/kknn/versions/1.3.1) needed for machine leraning functions.

```{r, warning=FALSE, message=FALSE}
library("tidyverse")
library('extrafont')
library('caret')
library('kknn')
library("randomForest")
```

## Machine learning data set

Let's load the `dfml.csv` data set. The clr coordinates are scaled to zero mean and unity variance.

```{r}
dfml = read_csv('output/dfml.csv')
dfml$Maturity5 = factor(dfml$Maturity5)
dfml$yieldClass = factor(dfml$yieldClass)
clr_no <- c("clr_N", "clr_P", "clr_K", "clr_Mg", "clr_Ca", "clr_Fv")
dfml.sc <- dfml # copy
dfml.sc[, clr_no] <- apply(dfml.sc[, clr_no], 2, scale) # scale predictors
```

## Machine learning

### Data train and test splits

We randomly split the data into a training set (75% of the data) used to fit the models, and a testing set (remaining 25%) used for models' evaluation.

```{r}
set.seed(8539)
split_index <- createDataPartition(dfml.sc$yieldClass, group = "Cultivar",
                                   p = 0.75, list = FALSE, times = 1)
train_df <- dfml.sc[split_index, ]
test_df <- dfml.sc[-split_index, ]
ml_col <- c(clr_no, "yieldClass")
```

With the `kknn` package, we must specify three parameters: `kmax` which is the number of neighbors to consider,  `distance` a distance parameter to specify (1 for the Mahattan distance and 2 for the Euclidean distance), and a `kernel` which is a function to measure the distance. A best method currently used to choose the right parameters consists in creating a parameter grid.

```{r}
(grid <-  expand.grid(kmax = c(7,9,12,15), distance = 1:2,
                      kernel = c("rectangular", "gaussian", "optimal")))
```

The metric of “Accuracy” is used to evaluate models quality. This is the ratio of the number of correctly predicted instances divided by the total number of instances in the dataset (e.g. 95% accurate). 

The accuracy of the models will be estimated using a `10-fold cross-validation (cv)` scheme. This will split the data set into 10 subsets of equal size. The models are built 10 times, each time leaving out one of the subsets from training and use it as the test set.

```{r}
control <- trainControl(method = "cv", number = 10)
metric <- "Accuracy"
```

### Building the Models

We reset the random number seed before each run to ensure that the evaluation of each algorithm is performed using exactly the same data splits. It ensures the results are directly comparable [Jason Brownlee](https://machinelearningmastery.com/machine-learning-in-r-step-by-step/).

```{r}
# a) Non-linear algorithm
## kNN
set.seed(7)
kknn_ <- train(yieldClass ~., data = train_df %>% select(ml_col), method = "kknn", 
                  metric = metric, trControl = control, tuneGrid = grid)
```

```{r}
# b) Advanced algorithms
## SVM
set.seed(7)
svm_ <- train(yieldClass ~., data = train_df %>% select(ml_col), method = "svmRadial", 
                 metric = metric, trControl = control)
```

```{r}
## Random Forest
set.seed(7)
rf_ <- train(yieldClass ~., data = train_df %>% select(ml_col), method = "rf", 
                metric = metric, trControl = control)
```

### Goodness of fit on training set

We assess the accuracy metric also during modeling (with train set) but the target metric is for the evaluation set. This chart sorts models.

```{r ml-traindotplot, fig.height = 3, fig.width = 5, out.width="100%", fig.align="center", warning=FALSE, message=FALSE, fig.cap="Comparison of models accuracies at training."}
# Summary results
results <- resamples(list(kknn_model = kknn_, 
                          svm_model = svm_, 
                          rf_model = rf_))
#summary(results) with dotplot()
dotplot(results)
```

This chunk also sorts models in a descending order using accuracies only, in a table.

```{r}
models_acc <- data.frame(Model = summary(results)$models,
                       Accuracy = c(confusionMatrix(train_df$yieldClass, predict(kknn_))$overall[1],
                                    confusionMatrix(train_df$yieldClass, predict(svm_))$overall[1],
                                    confusionMatrix(train_df$yieldClass, predict(rf_))$overall[1]))
models_acc[order(models_acc[,"Accuracy"], decreasing = TRUE), ]
```

The next one prints the best tuning parameters that maximizes model accuracy.

```{r}
data.frame(Model = summary(results)$models,
           param = c(kknn_$bestTune, 
                     svm_$bestTune, 
                     rf_$bestTune))
```

### Models' evaluation (test set)

Model evaluation is an integral part of the model development process. It helps to find the best model that represents our data and how well the chosen model will work in the future. The next chunk performs this computations and gives the sorted accuracy metrics.

```{r}
predicted_kknn_ <- predict(kknn_, test_df %>% select(ml_col))
predicted_svm_ <- predict(svm_, test_df %>% select(ml_col))
predicted_rf_ <- predict(rf_, test_df %>% select(ml_col))

#The best model
tests_acc <- data.frame(Model = summary(results)$models,
                        Accuracy_on_test = c(
                          confusionMatrix(test_df$yieldClass, predicted_kknn_)$overall[1],
                          confusionMatrix(test_df$yieldClass, predicted_svm_)$overall[1],
                          confusionMatrix(test_df$yieldClass, predicted_rf_)$overall[1]))
tests_acc[order(tests_acc[,"Accuracy_on_test"], decreasing = TRUE), ]
```

The k-nearest neighbours, the random forest and the support vector machine models returned similar predictive accuracies (_although slightly higher for the former_). 

## Variable importance estimation

The `varImp()` method is then used to estimate the variable importance, which is printed (summarized) and plotted. `varImp()` ranks features by importance.

```{r}
importance <- varImp(rf_, scale = FALSE) # scale between 1 to 100
print(importance)
```

```{r, var-imp-plot, fig.height = 2, fig.width = 4, out.width="100%", fig.align="center", warning=FALSE, message=FALSE, fig.cap="Importance of clr variables (effect) in the model."}
#tiff('images/var-imp-plot.tiff')
plot(importance, cex = 1.2, cex.lab = 2, cex.axis = 2, ylab = "variable", col = "black")
#dev.off()
```

## Make predictions on the test set with `kknn` model

We sort the predictive quality metrics by cultivar with `random forest` algorithm: model `rf_clr`, and tide the table.

```{r}
test_df$ypred = predicted_kknn_ # adds predictions to test set

cultivar_acc <- test_df %>%
    group_by(Cultivar) %>%
    do(Accuracy = as.numeric(confusionMatrix(.$yieldClass, .$ypred)$overall[1]),
       numb_samples = as.numeric(nrow(.)))

cultivar_acc$Accuracy <- round(unlist(cultivar_acc$Accuracy), 2)
cultivar_acc$numb_samples <- unlist(cultivar_acc$numb_samples)

data = data.frame(subset(cultivar_acc, Accuracy>0))
data[order(data[,"Accuracy"], decreasing=T), ]
```

The predictive accuracy is very high for some cultivars, but this result must be taken with care due to small size of samples often available. We plot accuracies for cultivars using ggplot2 functions.

```{r accuracy-cultivar, fig.height = 5, fig.width = 9, out.width="100%", echo = FALSE, fig.align="center", warning=FALSE, message=FALSE, fig.cap="Predictive accuracy for cultivars."}
ggplot(data, aes(x = reorder(Cultivar, Accuracy), y = Accuracy)) +
  geom_col(aes(fill = numb_samples)) +
  scale_fill_gradient(low = "grey50", high = "black") +
  labs(x = 'Cultivar', y = 'Accuracy', fill = "Number of\nsamples") +
  theme_bw() +
  theme(text = element_text(family = "Arial", face = "bold", size = 12)) +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5, size = 12))
ggsave("images/cultivar_accuracy.tiff", width = 9, height = 5)
```

## Comparison with non-informative classification

The non-informative classification consists of an equal distribution of 50% successful and 50% unsuccessful classification cases [(Swets J. A., 1988)](https://science.sciencemag.org/content/240/4857/1285). We run the Chi-square homogenity test to compare predictive accuracy of the __..kknn..__ model to this non-informative classification model. 

```{r}
cm <- confusionMatrix(predicted_kknn_, test_df$yieldClass) 
cm$table # confusion matrix
```

```{r}
# rf_clr_and_ionomicgroup model's classification
good_class <- cm$table[1,1]+cm$table[2,2] # HY or LY and correctly predicted
misclass <- cm$table[1,2]+cm$table[2,1]   # wrong classification
ml_class <- c(good_class, misclass)

# Non-informative model (nim)
total <- sum(cm$table) # total number of samples
good_nim <- 0.50 * total
misclass_nim <- 0.50 * total
non_inf_model <- c(good_nim, misclass_nim)

# Matrix for chisquare test
m <- rbind(ml_class, non_inf_model)
m

# chisq.test
khi2_test <- chisq.test(m)
khi2_test
```

The null hypothesis for a non-informative classification is rejected after the chi-square homogeneity test, the too small p-value denotes important difference between the models.

## Nutritionally balanced compositions

We consider as True Negatives (TN) specimens for this study, observations of the training data set having a high yield (HY) and correctly predicted with the `__k nearest neighbors__` model. 

Let's save prediction for training set.

```{r}
tr_pred <- predict(kknn_)#, train_df %>% select(ml_col))
train_df <- bind_cols(dfml[split_index, ], pred_yield = tr_pred)
```


```{r}
TNs = train_df[train_df$yieldClass == 'HY' & tr_pred == 'HY', ]
```

Then, we compute __clr centroids (means)__ for cultivars using True Negatives original (not scaled) clr values. These centroids (__S3 Table of the article__) could be used as provisional `clr mean norms` for cultivars. The standard deviations could also be computed. We considered all the TNs as norms. Cultivars Lamoka and Sifra had no true negative specimens.

```{r}
TNmeanNorms <- TNs %>%
  group_by(Cultivar) %>%
  select(clr_no) %>%
  summarise_all(list(mean))
TNmeanNorms
write_csv(TNmeanNorms, "output/provisional_meanNorms.csv")
```

We also save prediction for testing set.

```{r}
te_pred <- predict(kknn_, test_df %>% select(ml_col))
test_df <- bind_cols(dfml[-split_index, ], pred_yield = te_pred)
```

And finally backup for the next chapter.

```{r}
write_csv(train_df, "output/train_df.csv")
write_csv(test_df, "output/test_df.csv")
```
