==> 04_perturbation-concept.Rmd <==
# Ionome perturbation concept {#Chapter-Perturbation}

## Objective

***
The objective of this chapter is to show the user a visual example of perturbation effect in a Simplex, and to develop a mathematical workflow useful to adjust the ionome of potato crops for diagnostic purpose. 

> Perturbation in compositional space plays the same role as translation plays in real space. The assumption is that some natural processes in nature can be interpreted as a change from one composition `C1` to another `C2` through the application of a perturbation:

> p ⊕ C1 ===> C2.

> The `difference` between `a new observation`  and a closest healthy composition (closest true negative - TN) or `reference` composition can be back-transformed to the compositional space. The resulted vector is the `perturbation vector`. 

Theoretically, a misbalanced composition could be balanced (translated into a healthy zone) using a perturbation operation. Using this concept, ionome of a new cultivar could be assigned to the cultivar sharing similar leaf composition, and where nutrient requirements have been already documented by fertilizer trials.

We used the __testing set__ to display the effect of a perturbation on the whole simplex. We selected two elements (N and P) and simulated an increase of their initial (observed) clr values by 20% (theoretically). The observed (observation) and new clr vector (perturbation) were back transformed to N, P, K, Ca, Mg and Fv compositional space for comparison. 

Secondly, the procedure used to rebalance a misbalanced composition is decribed. As explained at the end of the Chapter \@ref(Chapter-Modeling), we consider as True Negatives (TN) specimens (or healthy points) for this study, observations of the __training set__ having a high yield (HY) and correctly predicted by the __k nearest neighbors__ model.

***

## Data set and useful libraries

We need package [compositions](https://www.rdocumentation.org/packages/compositions/versions/1.40-2) for further clr back-transformation to compositional space. The package [reshape](https://cran.r-project.org/web/packages/reshape/index.html) will be used to melt an intermediate data frame. 

```{r, warning=FALSE, message=FALSE}
library("tidyverse")
library("extrafont")
library('compositions')
library("reshape")
```

The previous `train_df` and `test_df` are loaded.

```{r}
train_df = read_csv("output/train_df.csv")
test_df = read_csv("output/test_df.csv")
TNs = train_df %>% filter(yieldClass == 'HY' & pred_yield == 'HY')
clr_no = c("clr_N", "clr_P", "clr_K", "clr_Ca", "clr_Mg", "clr_Fv")
```

Filtrer train_df et test_df pour ne conserver que les observations ayant les cultivars correspondant dans les vrais négatifs, et seulement les déséquilibrés.

```{r}
train_df <- train_df %>% filter(Cultivar %in% unique(TNs$Cultivar))
test_df <- test_df %>% filter(Cultivar %in% unique(TNs$Cultivar))
```

## Euclidean distance from nutritionally balanced compositions

The chunk below activates the custom function used to compute Euclidean distance.

```{r}
eucl_dist_f <- function(x, y) {
    sqrt(sum((x-y)^2))
}
```

For each imbalanced composition, we use the next loop to compute all the euclidean distances between all the compositions in "TNs" of the corresponding cultivar. The computation is possible even if the cultivar is unknown, the loop must just be updated. Here, the loop returns the smallest Euclidean distance stored in `debal` vector.

For train_df:

```{r}
debal <- c()
debal_index <- c()
for (i in 1:nrow(train_df)) {
    clr_i <- as.numeric(train_df[i, clr_no])
    TNs_target <- TNs %>% filter(Cultivar == train_df$Cultivar[i]) %>% select(clr_no)
    eucl_dist <- apply(TNs_target, 1, function(x) eucl_dist_f(x = x, y = clr_i))
    debal_index[i] <- which.min(eucl_dist)
    debal[i] <- eucl_dist[debal_index[i]]
}
train_df$debal <- debal
train_df <- train_df %>% filter(debal != 0)
train_df %>% glimpse()
```

For test_df:

```{r}
debal <- c()
debal_index <- c()
for (i in 1:nrow(test_df)) {
    clr_i <- as.numeric(test_df[i, clr_no])
    TNs_target <- TNs %>% filter(Cultivar == test_df$Cultivar[i]) %>% select(clr_no)
    eucl_dist <- apply(TNs_target, 1, function(x) eucl_dist_f(x = x, y = clr_i))
    debal_index[i] <- which.min(eucl_dist)
    debal[i] <- eucl_dist[debal_index[i]]
}
test_df$debal <- debal
test_df <- test_df %>% filter(debal != 0)
test_df %>% glimpse()
```

## Perturbation effect of some elements on the whole

This subsection illustrates the principle that strictly positive data constrained to some whole are inherently related to each other. Changing a proportion (so, perturbation on some proportion(s)) inherently affects at least another proportion, because such data convey only relative information [(Aitchison, 1982)](https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.2517-6161.1982.tb01195.x). 

`leaf_clr_o` stands for original clr values of the tidded test set `test_df`.

```{r}
# Compute (or select here) the clrs
leaf_clr_o <- test_df %>% select(clr_no)
#leaf_clr_o <- test_df %>% 
#  filter(debal >= quantile(train_df$debal, p = .75)) %>% 
#  select(clr_no)
summary(leaf_clr_o)
```

Let's perturb the original clr values for `N` and `P`.

```{r}
# Perturb the original clrs
pert_col <- c(1, 2) # the column indices which is perturbed: clr_N and clr_K respectively
perturbation <- c(0.2, 0.2) # the amount added to the clr of the pert_col, same lenght as pert_col
leaf_clr_f <- leaf_clr_o
for (i in seq_along(pert_col)) {
  leaf_clr_f[, pert_col[i]] <- leaf_clr_f[, pert_col[i]] * (1 + perturbation[i])
}
summary(leaf_clr_f)
```

The next one transforms clrs (original and perturbed clrs) back to compositions.

```{r}
# From clrs to compositions
leaf_o <- apply(leaf_clr_o, 1, function(x) exp(x) / sum(exp(x))) %>% t()
leaf_f <- apply(leaf_clr_f, 1, function(x) exp(x) / sum(exp(x))) %>% t()
```

Then, we plot the original and perturbed ionomes to check a general tendency. `Observation` column plots the original "N", "P", "K", "Ca", "Mg" and "Fv" compositions, `Perturbation` represents new compositions after perturbation and `Difference` column stands for perturbation occured in the Observation to yied new compositions. Data are tidded before.

```{r}
rshleaf_o <- melt(data.frame(leaf_o)) %>% mutate(vector = rep("Observation", nrow(.)))
rshleaf_f <- melt(data.frame(leaf_f)) %>% mutate(vector = rep("Perturbation", nrow(.)))
rshdf <- bind_rows(rshleaf_o, rshleaf_f)
rshdf$is_perturbed <- ifelse(rshdf$variable %in% colnames(leaf_o[, pert_col]), 
                          "Perturbed", "Not perturbed")
```

```{r}
rshdf$variable <- sub(pattern = "clr_", replacement = "", x = rshdf$variable, fixed = TRUE) %>% 
  fct_relevel("N", "P", "K", "Ca", "Mg", "Fv")
```

```{r, wraped-perturb-boxplots, fig.height = 8, fig.width = 8, out.width = "100%", fig.align = "center", warning = FALSE, message = FALSE, echo = FALSE, fig.cap = "Perturbation effect of some elements on the whole."}
axis_titles <- paste("Proportion of", levels(rshdf$variable))
names(axis_titles) <- levels(rshdf$variable)
ggplot(data = rshdf, aes(x = vector, y = value)) +
  facet_wrap(. ~ variable, scales = "free", ncol = 3,
             strip.position = "left", 
             labeller = as_labeller(axis_titles)) +
  geom_boxplot(aes(fill = is_perturbed), outlier.size = 0.1) +
  labs(x = "", y = "",
       fill = "Was the associated clr perturbed?") +
  scale_fill_manual(values=c("white", "grey80")) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        strip.background = element_blank(),
        strip.placement = "outside",
        legend.position="bottom",
        panel.spacing = unit(2.5, "lines")) +
  theme(text = element_text(family = "Arial", face = "bold", size = 12))
ggsave("images/wraped_perturb_boxplots.tiff", height = 8, width = 8, dpi = 300)
```

All the components change when the clr of a single component is offset. The components whose clr has been perturbed obviously change the most (2-Perturbation). The component whose clr is the highest (generally Fv) compensate most of the perturbation. Although P clr values have been increased, P proportion decreased globally for the new equilibrium of the simplex.

## Rebalancing a misbalanced sample by perturbation

Let's suppose that we got this point selected at random in `imbalanced or misbalanced specimens`.

```{r}
set.seed(92559) 
imbalanced <- train_df[sample(nrow(train_df), 1), ]
t(imbalanced)
```

Or even, we could rather use the most imbalanced occurrence, why not !

```{r}
imbalanced <- train_df[which.max(train_df$debal), ]
misbalanced <- imbalanced # copy
t(misbalanced)
```

How could we rebalance it? The first step is to find the closest balanced point in the TNs of the corresponding cultivar. Let's re-compute its Euclidean distances from TNs and identify the TNs' sample from which the distance is minimum.

```{r}
misbalanced <- misbalanced[clr_no]
eucl_dist_misbal <- apply(TNs %>% filter(Cultivar == imbalanced$Cultivar) %>% select(clr_no), 
                          1, function(x) eucl_dist_f(x = x, y = misbalanced))

index_misbal <- which.min(t(data.frame(eucl_dist_misbal))) 
index_misbal # return the index of the sample
```

The closest healthy sample is the one which index is `r index_misbal` in TNs charing the same cultivar with the new sample. Using this index we could refind the minimum imbalance index value computed.

```{r}
(misbal <- eucl_dist_misbal[index_misbal])
```

The Euclidean distance matches with the corresponding `debal` value: `imbalanced$debal[1]` = `r imbalanced$debal[1]`. The `closest point in the TNs` subset is this one:

```{r}
target_TNs <- TNs %>% 
  filter(Cultivar == imbalanced$Cultivar)
closest <- target_TNs[index_misbal, ]
t(closest)
```

Note that `Cultivar` of the misbalanced and the closest healthy composition are the same. We compute the clr difference between the closest and the misbalanced points. 

```{r}
closest = closest[clr_no]
clr_diff = closest - misbalanced
t(clr_diff)
```

The perturbation vector is that clr difference back-transformed to leaf compositional space.

```{r}
comp_names <- c("N", "P", "K", "Ca", "Mg", "Fv")
perturbation_vector <- clrInv(clr_diff)
names(perturbation_vector) <- comp_names
t(perturbation_vector)
```

Next, we should compute the corresponding compositions of the clr coordinates of the misbalanced point, as well as the closest TN point. The vectors could be gathered in a table made up of perturbation vector, misbalanced composition and the closest reference sample (pmc).

```{r}
misbal_comp <- clrInv(misbalanced)
names(misbal_comp) <- comp_names

closest_comp <- clrInv(closest)
names(closest_comp) <- comp_names

pmc = rbind(perturbation_vector, misbal_comp, closest_comp)
rownames(pmc) = c("perturbation_vector","misbal_comp","closest_comp")
pmc
```

We could even check that the simplex is closed to 1 for each vector.

```{r}
sum(perturbation_vector); sum(misbal_comp); sum(closest_comp)
```

The closest composition minus the misbalanced composition should return the perturbation vector.

```{r}
print(closest_comp - misbal_comp)
print(perturbation_vector) # for comparison
```

Or even, perturb the misbalanced point by the perturbation vector, you should obtain the closest TN point:

```{r}
print(misbal_comp + perturbation_vector) # perturbation
print(closest_comp)                      # for comparison
```

So, the assumption is true. The next codes show the concept using dots plots and histograms for each vector. A data frame is tidded for ggplot. Visualization is better with histograms.

```{r}
df <- data.frame(rbind(misbalanced, closest, clr_diff),
                 vectors = factor(c("Observation", "Reference", "Perturbation")))
df$vectors <- df$vectors %>% fct_relevel("Observation", "Reference", "Perturbation")
dfreshape <- melt(df) # reshapes df for ggplot
dfreshape <- melt(df)
dfreshape$variable <- sub(pattern = "clr_", replacement = "", x = dfreshape$variable, fixed = TRUE) %>% 
  fct_relevel("N", "P", "K", "Ca", "Mg", "Fv")
```

```{r hi-perturb-dotplot, fig.height = 4, fig.width = 10, out.width = "100%", fig.align = "center", warning = FALSE, message = FALSE, fig.cap = "Perturbation vector computation example dotplot using the most imbalanced foliar sample."}
ggplot(data = dfreshape, aes(x = value, y = vectors, colour = vectors)) +
    geom_point() +
    facet_wrap(~ variable, scales = "free_x") +
    labs(x='Nutrient clr coordinate', y ='') +
    theme(text=element_text(family="Arial", face="bold", size=12)) +
    theme_bw()
ggsave("images/perturb_dotplot.tiff",  width = 7, height = 3)
```

```{r hi-perturb-barplot, fig.height = 4, fig.width = 6, out.width="100%", fig.align="center", warning=FALSE, message=FALSE, fig.cap="Perturbation vector computation example barplot using the most imbalanced foliar sample."}
g1 <- ggplot(data = dfreshape, aes(x = variable, y = value, fill = vectors)) +
    geom_bar(aes(fill = vectors), stat = "identity", position = position_dodge()) +
    coord_flip() + theme_bw() +
    ylab("Nutrients clr coordinates") + xlab("Diagnostic nutrients") +
    theme(legend.title = element_blank()) +
    theme(text = element_text(family = "Arial", face = "bold", size = 12))
g1 + scale_fill_discrete(breaks = c("Observation","Reference","Perturbation")) +
    scale_fill_manual(values=c("grey50", "black", "grey80"))

ggsave("images/perturb_barplot.tiff",  width = 6, height = 4)
```
